git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/interpolation-model-checking@38916 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/core/counterexample/Address.java
@@ -115,7 +115,7 @@ public abstract class Address {
       long value = ((Integer) pAddress).longValue();
       return new ConcreteAddress(BigInteger.valueOf(value));
     } else if (pAddress instanceof Long) {
-      long value = ((Long) pAddress).longValue();
+      long value = ((Long) pAddress);
       return new ConcreteAddress(BigInteger.valueOf(value));
     } else if (pAddress instanceof Short) {
       long value = ((Short) pAddress).longValue();
